move interfaces out of model
