+++ packages/backend/constants/hubspot.ts
@@ -1,0 +1,506 @@
+export interface HubspotContact {
+    company_size: string;
+    date_of_birth: string;
+    degree: string;
+    field_of_study: string;
+    first_deal_created_date: Date;
+    gender: string;
+    graduation_date: string;
+    hs_all_assigned_business_unit_ids: number;
+    hs_analytics_first_touch_converting_campaign: string;
+    hs_analytics_last_touch_converting_campaign: string;
+    hs_avatar_filemanager_key: string;
+    hs_buying_role: number;
+    hs_clicked_linkedin_ad: number;
+    hs_content_membership_email: string;
+    hs_content_membership_email_confirmed: boolean;
+    hs_content_membership_follow_up_enqueued_at: Date;
+    hs_content_membership_notes: string;
+    hs_content_membership_registered_at: Date;
+    hs_content_membership_registration_domain_sent_to: string;
+    hs_content_membership_registration_email_sent_at: Date;
+    hs_content_membership_status: number;
+    hs_conversations_visitor_email: string;
+    hs_count_is_unworked: number;
+    hs_count_is_worked: number;
+    hs_created_by_conversations: boolean;
+    hs_created_by_user_id: number;
+    hs_createdate: Date;
+    hs_date_entered_customer: Date;
+    hs_date_entered_evangelist: Date;
+    hs_date_entered_lead: Date;
+    hs_date_entered_marketingqualifiedlead: Date;
+    hs_date_entered_opportunity: Date;
+    hs_date_entered_other: Date;
+    hs_date_entered_salesqualifiedlead: Date;
+    hs_date_entered_subscriber: Date;
+    hs_date_exited_customer: Date;
+    hs_date_exited_evangelist: Date;
+    hs_date_exited_lead: Date;
+    hs_date_exited_marketingqualifiedlead: Date;
+    hs_date_exited_opportunity: Date;
+    hs_date_exited_other: Date;
+    hs_date_exited_salesqualifiedlead: Date;
+    hs_date_exited_subscriber: Date;
+    hs_document_last_revisited: Date;
+    hs_email_bad_address: boolean;
+    hs_email_customer_quarantined_reason: number;
+    hs_email_hard_bounce_reason: string;
+    hs_email_hard_bounce_reason_number: number;
+    hs_email_quarantined: boolean;
+    hs_email_quarantined_reason: number;
+    hs_email_recipient_fatigue_recovery_time: Date;
+    hs_email_sends_since_last_engagement: number;
+    hs_emailconfirmationstatus: number;
+    hs_facebook_ad_clicked: boolean;
+    hs_facebook_click_id: string;
+    hs_facebookid: string;
+    hs_feedback_last_nps_follow_up: string;
+    hs_feedback_last_nps_rating: number;
+    hs_feedback_last_survey_date: Date;
+    hs_feedback_show_nps_web_survey: boolean;
+    hs_first_subscription_create_date: Date;
+    hs_google_click_id: string;
+    hs_googleplusid: string;
+    hs_has_active_subscription: number;
+    hs_ip_timezone: string;
+    hs_is_unworked: boolean;
+    hs_last_sales_activity_date: Date;
+    hs_last_sales_activity_timestamp: Date;
+    hs_last_sales_activity_type: number;
+    hs_lastmodifieddate: Date;
+    hs_latest_sequence_enrolled: number;
+    hs_latest_sequence_enrolled_date: Date;
+    hs_latest_sequence_finished_date: Date;
+    hs_latest_sequence_unenrolled_date: Date;
+    hs_latest_source_timestamp: Date;
+    hs_latest_subscription_create_date: Date;
+    hs_lead_status: number;
+    hs_legal_basis: number;
+    hs_linkedin_ad_clicked: number;
+    hs_linkedinid: string;
+    hs_marketable_reason_id: string;
+    hs_marketable_reason_type: number;
+    hs_marketable_status: number;
+    hs_marketable_until_renewal: number;
+    hs_merged_object_ids: number;
+    hs_object_id: number;
+    hs_pinned_engagement_id: number;
+    hs_pipeline: number;
+    hs_predictivecontactscore_v2: number;
+    hs_predictivescoringtier: number;
+    hs_read_only: boolean;
+    hs_sa_first_engagement_date: Date;
+    hs_sales_email_last_clicked: Date;
+    hs_sales_email_last_opened: Date;
+    hs_sequences_actively_enrolled_count: number;
+    hs_sequences_enrolled_count: number;
+    hs_testpurge: string;
+    hs_testrollback: string;
+    hs_time_between_contact_creation_and_deal_close: number;
+    hs_time_between_contact_creation_and_deal_creation: number;
+    hs_time_in_customer: number;
+    hs_time_in_evangelist: number;
+    hs_time_in_lead: number;
+    hs_time_in_marketingqualifiedlead: number;
+    hs_time_in_opportunity: number;
+    hs_time_in_other: number;
+    hs_time_in_salesqualifiedlead: number;
+    hs_time_in_subscriber: number;
+    hs_time_to_move_from_lead_to_customer: number;
+    hs_time_to_move_from_marketingqualifiedlead_to_customer: number;
+    hs_time_to_move_from_opportunity_to_customer: number;
+    hs_time_to_move_from_salesqualifiedlead_to_customer: number;
+    hs_time_to_move_from_subscriber_to_customer: number;
+    hs_timezone: number;
+    hs_twitterid: string;
+    hs_unique_creation_key: string;
+    hs_updated_by_user_id: number;
+    hs_user_ids_of_all_notification_followers: number;
+    hs_user_ids_of_all_notification_unfollowers: number;
+    hs_user_ids_of_all_owners: number;
+    hs_whatsapp_phone_number: string;
+    hubspot_owner_assigneddate: Date;
+    ip_city: string;
+    ip_country: string;
+    ip_country_code: string;
+    ip_latlon: string;
+    ip_state: string;
+    ip_state_code: string;
+    ip_zipcode: string;
+    job_function: string;
+    lastmodifieddate: Date;
+    marital_status: string;
+    military_status: string;
+    num_associated_deals: number;
+    recent_deal_amount: number;
+    recent_deal_close_date: Date;
+    relationship_status: string;
+    school: string;
+    seniority: string;
+    start_date: string;
+    total_revenue: number;
+    work_email: string;
+    firstname: string;
+    hs_analytics_first_url: string;
+    hs_email_delivered: number;
+    hs_email_optout_78547538: number;
+    twitterhandle: string;
+    currentlyinworkflow: number;
+    followercount: number;
+    hs_analytics_last_url: string;
+    hs_email_open: number;
+    lastname: string;
+    hs_analytics_num_page_views: number;
+    hs_email_click: number;
+    salutation: string;
+    twitterprofilephoto: string;
+    email: string;
+    hs_analytics_num_visits: number;
+    hs_email_bounce: number;
+    hs_persona: number;
+    hs_social_last_engagement: Date;
+    hs_analytics_num_event_completions: number;
+    hs_email_optout: boolean;
+    hs_social_twitter_clicks: number;
+    mobilephone: string;
+    phone: string;
+    fax: string;
+    hs_analytics_first_timestamp: Date;
+    hs_email_last_email_name: string;
+    hs_email_last_send_date: Date;
+    hs_social_facebook_clicks: number;
+    address: string;
+    engagements_last_meeting_booked: Date;
+    engagements_last_meeting_booked_campaign: string;
+    engagements_last_meeting_booked_medium: string;
+    engagements_last_meeting_booked_source: string;
+    hs_analytics_first_visit_timestamp: Date;
+    hs_email_last_open_date: Date;
+    hs_latest_meeting_activity: Date;
+    hs_sales_email_last_replied: Date;
+    hs_social_linkedin_clicks: number;
+    hubspot_owner_id: number;
+    notes_last_contacted: Date;
+    notes_last_updated: Date;
+    notes_next_activity_date: Date;
+    num_contacted_notes: number;
+    num_notes: number;
+    owneremail: string;
+    ownername: string;
+    surveymonkeyeventlastupdated: number;
+    webinareventlastupdated: number;
+    city: string;
+    hs_analytics_last_timestamp: Date;
+    hs_email_last_click_date: Date;
+    hs_social_google_plus_clicks: number;
+    hubspot_team_id: number;
+    linkedinbio: string;
+    twitterbio: string;
+    hs_all_owner_ids: number;
+    hs_analytics_last_visit_timestamp: Date;
+    hs_email_first_send_date: Date;
+    hs_social_num_broadcast_clicks: number;
+    state: string;
+    hs_all_team_ids: number;
+    hs_analytics_source: number;
+    hs_email_first_open_date: Date;
+    hs_latest_source: number;
+    zip: string;
+    country: string;
+    hs_all_accessible_team_ids: number;
+    hs_analytics_source_data_1: string;
+    hs_email_first_click_date: Date;
+    hs_latest_source_data_1: string;
+    linkedinconnections: number;
+    hs_analytics_source_data_2: string;
+    hs_email_is_ineligible: boolean;
+    hs_language: number;
+    hs_latest_source_data_2: string;
+    kloutscoregeneral: number;
+    hs_analytics_first_referrer: string;
+    hs_email_first_reply_date: Date;
+    jobtitle: string;
+    photo: string;
+    hs_analytics_last_referrer: string;
+    hs_email_last_reply_date: Date;
+    message: string;
+    closedate: Date;
+    hs_analytics_average_page_views: number;
+    hs_email_replied: number;
+    hs_analytics_revenue: number;
+    hs_lifecyclestage_lead_date: Date;
+    hs_lifecyclestage_marketingqualifiedlead_date: Date;
+    hs_lifecyclestage_opportunity_date: Date;
+    lifecyclestage: number;
+    hs_lifecyclestage_salesqualifiedlead_date: Date;
+    createdate: Date;
+    hs_lifecyclestage_evangelist_date: Date;
+    hs_lifecyclestage_customer_date: Date;
+    hubspotscore: number;
+    company: string;
+    hs_lifecyclestage_subscriber_date: Date;
+    hs_lifecyclestage_other_date: Date;
+    website: string;
+    numemployees: number;
+    annualrevenue: string;
+    industry: string;
+    associatedcompanyid: number;
+    associatedcompanylastupdated: number;
+    hs_predictivecontactscorebucket: number;
+    hs_predictivecontactscore: number;
+}
+
+export interface HubspotLead {
+    company_size: string;
+    date_of_birth: string;
+    degree: string;
+    field_of_study: string;
+    first_deal_created_date: Date;
+    gender: string;
+    graduation_date: string;
+    hs_all_assigned_business_unit_ids: number;
+    hs_analytics_first_touch_converting_campaign: string;
+    hs_analytics_last_touch_converting_campaign: string;
+    hs_avatar_filemanager_key: string;
+    hs_buying_role: number;
+    hs_clicked_linkedin_ad: number;
+    hs_content_membership_email: string;
+    hs_content_membership_email_confirmed: boolean;
+    hs_content_membership_follow_up_enqueued_at: Date;
+    hs_content_membership_notes: string;
+    hs_content_membership_registered_at: Date;
+    hs_content_membership_registration_domain_sent_to: string;
+    hs_content_membership_registration_email_sent_at: Date;
+    hs_content_membership_status: number;
+    hs_conversations_visitor_email: string;
+    hs_count_is_unworked: number;
+    hs_count_is_worked: number;
+    hs_created_by_conversations: boolean;
+    hs_created_by_user_id: number;
+    hs_createdate: Date;
+    hs_date_entered_customer: Date;
+    hs_date_entered_evangelist: Date;
+    hs_date_entered_lead: Date;
+    hs_date_entered_marketingqualifiedlead: Date;
+    hs_date_entered_opportunity: Date;
+    hs_date_entered_other: Date;
+    hs_date_entered_salesqualifiedlead: Date;
+    hs_date_entered_subscriber: Date;
+    hs_date_exited_customer: Date;
+    hs_date_exited_evangelist: Date;
+    hs_date_exited_lead: Date;
+    hs_date_exited_marketingqualifiedlead: Date;
+    hs_date_exited_opportunity: Date;
+    hs_date_exited_other: Date;
+    hs_date_exited_salesqualifiedlead: Date;
+    hs_date_exited_subscriber: Date;
+    hs_document_last_revisited: Date;
+    hs_email_bad_address: boolean;
+    hs_email_customer_quarantined_reason: number;
+    hs_email_hard_bounce_reason: string;
+    hs_email_hard_bounce_reason_number: number;
+    hs_email_quarantined: boolean;
+    hs_email_quarantined_reason: number;
+    hs_email_recipient_fatigue_recovery_time: Date;
+    hs_email_sends_since_last_engagement: number;
+    hs_emailconfirmationstatus: number;
+    hs_facebook_ad_clicked: boolean;
+    hs_facebook_click_id: string;
+    hs_facebookid: string;
+    hs_feedback_last_nps_follow_up: string;
+    hs_feedback_last_nps_rating: number;
+    hs_feedback_last_survey_date: Date;
+    hs_feedback_show_nps_web_survey: boolean;
+    hs_first_subscription_create_date: Date;
+    hs_google_click_id: string;
+    hs_googleplusid: string;
+    hs_has_active_subscription: number;
+    hs_ip_timezone: string;
+    hs_is_unworked: boolean;
+    hs_last_sales_activity_date: Date;
+    hs_last_sales_activity_timestamp: Date;
+    hs_last_sales_activity_type: number;
+    hs_lastmodifieddate: Date;
+    hs_latest_sequence_enrolled: number;
+    hs_latest_sequence_enrolled_date: Date;
+    hs_latest_sequence_finished_date: Date;
+    hs_latest_sequence_unenrolled_date: Date;
+    hs_latest_source_timestamp: Date;
+    hs_latest_subscription_create_date: Date;
+    hs_lead_status: number;
+    hs_lead_source: string;
+    hs_legal_basis: number;
+    hs_linkedin_ad_clicked: number;
+    hs_linkedinid: string;
+    hs_marketable_reason_id: string;
+    hs_marketable_reason_type: number;
+    hs_marketable_status: number;
+    hs_marketable_until_renewal: number;
+    hs_merged_object_ids: number;
+    hs_object_id: number;
+    hs_pinned_engagement_id: number;
+    hs_pipeline: number;
+    hs_predictivecontactscore_v2: number;
+    hs_predictivescoringtier: number;
+    hs_read_only: boolean;
+    hs_sa_first_engagement_date: Date;
+    hs_sales_email_last_clicked: Date;
+    hs_sales_email_last_opened: Date;
+    hs_sequences_actively_enrolled_count: number;
+    hs_sequences_enrolled_count: number;
+    hs_testpurge: string;
+    hs_testrollback: string;
+    hs_time_between_contact_creation_and_deal_close: number;
+    hs_time_between_contact_creation_and_deal_creation: number;
+    hs_time_in_customer: number;
+    hs_time_in_evangelist: number;
+    hs_time_in_lead: number;
+    hs_time_in_marketingqualifiedlead: number;
+    hs_time_in_opportunity: number;
+    hs_time_in_other: number;
+    hs_time_in_salesqualifiedlead: number;
+    hs_time_in_subscriber: number;
+    hs_time_to_move_from_lead_to_customer: number;
+    hs_time_to_move_from_marketingqualifiedlead_to_customer: number;
+    hs_time_to_move_from_opportunity_to_customer: number;
+    hs_time_to_move_from_salesqualifiedlead_to_customer: number;
+    hs_time_to_move_from_subscriber_to_customer: number;
+    hs_timezone: number;
+    hs_twitterid: string;
+    hs_unique_creation_key: string;
+    hs_updated_by_user_id: number;
+    hs_user_ids_of_all_notification_followers: number;
+    hs_user_ids_of_all_notification_unfollowers: number;
+    hs_user_ids_of_all_owners: number;
+    hs_whatsapp_phone_number: string;
+    hubspot_owner_assigneddate: Date;
+    ip_city: string;
+    ip_country: string;
+    ip_country_code: string;
+    ip_latlon: string;
+    ip_state: string;
+    ip_state_code: string;
+    ip_zipcode: string;
+    job_function: string;
+    lastmodifieddate: Date;
+    marital_status: string;
+    military_status: string;
+    num_associated_deals: number;
+    recent_deal_amount: number;
+    recent_deal_close_date: Date;
+    relationship_status: string;
+    school: string;
+    seniority: string;
+    start_date: string;
+    total_revenue: number;
+    work_email: string;
+    firstname: string;
+    hs_analytics_first_url: string;
+    hs_email_delivered: number;
+    hs_email_optout_78547538: number;
+    twitterhandle: string;
+    currentlyinworkflow: number;
+    followercount: number;
+    hs_analytics_last_url: string;
+    hs_email_open: number;
+    lastname: string;
+    hs_analytics_num_page_views: number;
+    hs_email_click: number;
+    salutation: string;
+    twitterprofilephoto: string;
+    email: string;
+    hs_analytics_num_visits: number;
+    hs_email_bounce: number;
+    hs_persona: number;
+    hs_social_last_engagement: Date;
+    hs_analytics_num_event_completions: number;
+    hs_email_optout: boolean;
+    hs_social_twitter_clicks: number;
+    mobilephone: string;
+    phone: string;
+    fax: string;
+    hs_analytics_first_timestamp: Date;
+    hs_email_last_email_name: string;
+    hs_email_last_send_date: Date;
+    hs_social_facebook_clicks: number;
+    address: string;
+    engagements_last_meeting_booked: Date;
+    engagements_last_meeting_booked_campaign: string;
+    engagements_last_meeting_booked_medium: string;
+    engagements_last_meeting_booked_source: string;
+    hs_analytics_first_visit_timestamp: Date;
+    hs_email_last_open_date: Date;
+    hs_latest_meeting_activity: Date;
+    hs_sales_email_last_replied: Date;
+    hs_social_linkedin_clicks: number;
+    hubspot_owner_id: number;
+    notes_last_contacted: Date;
+    notes_last_updated: Date;
+    notes_next_activity_date: Date;
+    num_contacted_notes: number;
+    num_notes: number;
+    owneremail: string;
+    ownername: string;
+    surveymonkeyeventlastupdated: number;
+    webinareventlastupdated: number;
+    city?: string;
+    hs_analytics_last_timestamp: Date;
+    hs_email_last_click_date: Date;
+    hs_social_google_plus_clicks: number;
+    hubspot_team_id: number;
+    linkedinbio: string;
+    twitterbio: string;
+    hs_all_owner_ids: number;
+    hs_analytics_last_visit_timestamp: Date;
+    hs_email_first_send_date: Date;
+    hs_social_num_broadcast_clicks: number;
+    state?: string;
+    hs_all_team_ids: number;
+    hs_analytics_source: number;
+    hs_email_first_open_date: Date;
+    hs_latest_source: number;
+    zip?: string;
+    country?: string;
+    hs_all_accessible_team_ids: number;
+    hs_analytics_source_data_1: string;
+    hs_email_first_click_date: Date;
+    hs_latest_source_data_1: string;
+    linkedinconnections: number;
+    hs_analytics_source_data_2: string;
+    hs_email_is_ineligible: boolean;
+    hs_language: number;
+    hs_latest_source_data_2: string;
+    kloutscoregeneral: number;
+    hs_analytics_first_referrer: string;
+    hs_email_first_reply_date: Date;
+    jobtitle: string;
+    photo: string;
+    hs_analytics_last_referrer: string;
+    hs_email_last_reply_date: Date;
+    message: string;
+    closedate: Date;
+    hs_analytics_average_page_views: number;
+    hs_email_replied: number;
+    hs_analytics_revenue: number;
+    hs_lifecyclestage_lead_date: Date;
+    hs_lifecyclestage_marketingqualifiedlead_date: Date;
+    hs_lifecyclestage_opportunity_date: Date;
+    lifecyclestage: number;
+    hs_lifecyclestage_salesqualifiedlead_date: Date;
+    createdate: Date;
+    hs_lifecyclestage_evangelist_date: Date;
+    hs_lifecyclestage_customer_date: Date;
+    hubspotscore: number;
+    company: string;
+    hs_lifecyclestage_subscriber_date: Date;
+    hs_lifecyclestage_other_date: Date;
+    website?: string;
+    numemployees: number;
+    annualrevenue: number;
+    industry: string;
+    associatedcompanyid: number;
+    associatedcompanylastupdated: number;
+    hs_predictivecontactscorebucket: number;
+    hs_predictivecontactscore: number;
+}
